+++ BACH/utils/argsUtil.h
@@ -42,7 +42,7 @@ struct Arguments {
   cl_int kernelOrder = 2;
 
   std::vector<cl_int> dg = {6, 4, 2};  // ngauss = length of dg
-  std::vector<double> bg = {
+  std::vector<float> bg = {
       (1.0 / (2.0 * 0.7 * 0.7)),
       (1.0 / (2.0 * 1.5 * 1.5)),
       (1.0 / (2.0 * 3.0 * 3.0)),
